requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-import-export==2.8.0
django-phonenumber-field==6.1.0
et-xmlfile==1.1.0
future==0.18.2
importlib-metadata==4.11.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.12.48
Pillow==9.1.1
pkg_resources==0.0.0
pystache==0.6.0
pytz==2022.1
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static==1.1.1
tablib==3.2.1
typing_extensions==4.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
